Updated README

+++ src/main/scala/com/group11/hw3/ChordMain.scala
@@ -136,7 +136,7 @@ object ChordMain {
     }
 
     Thread.sleep(1000)
-    val hopsPerReq = server.hopsPerReq
+    val hopsPerReq = server.hopsPerReq.toDouble
     var reqPerNode = 0.toDouble
     for (node <- chordNodesId) {
       val future = (chordShardRegion ? EntityEnvelope(node, GetStats())).mapTo[Int]
